# docs/swagger.yaml
definitions:
  models.Date:
    properties:
      time.Time:
        type: string
    type: object
  models.Project:
    properties:
      createdAt:
        type: string
      description:
        type: string
      endDate:
        $ref: '#/definitions/models.Date'
      id:
        type: string
      lastModified:
        type: string
      name:
        type: string
      startDate:
        $ref: '#/definitions/models.Date'
      status:
        type: string
      workspace:
        $ref: '#/definitions/models.Workspace'
    type: object
  models.Task:
    properties:
      createdAt:
        type: string
      description:
        type: string
      due:
        type: string
      id:
        type: string
      lastModified:
        type: string
      priority:
        $ref: '#/definitions/models.TaskPriority'
      project:
        $ref: '#/definitions/models.Project'
      status:
        $ref: '#/definitions/models.TaskStatus'
      title:
        type: string
    type: object
  models.TaskPriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
  models.TaskStatus:
    enum:
    - todo
    - started
    - complete
    type: string
    x-enum-varnames:
    - StatusTodo
    - StatusInProgress
    - StatusDone
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastModified:
        type: string
      name:
        type: string
      profilePhoto:
        type: string
      role:
        type: string
      verified:
        type: boolean
    type: object
  models.Workspace:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      lastModified:
        type: string
      name:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
info:
  contact:
    email: support@hazel.local
    name: API Support
    url: https://github.com/luckmade/hazel
  description: This is the backend API for the Hazel project management application.
  title: Hazel Project Management API
  version: "1.0"
paths:
...
  /projects/{id}/tasks:
    get:
      description: Get all tasks for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get project tasks
      tags:
      - tasks
  /tasks:
    post:
      consumes:
      - application/json
      description: Create a new task in a project
      parameters:
      - description: Task info
        in: body
        name: task
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get task
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Update task details
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task update info
        in: body
        name: task
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
  /tasks/{id}/assignments:
    get:
      description: Get all users assigned to a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get assigned users
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Assign a task to a user
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment info
        in: body
        name: assignment
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign task to user
      tags:
      - tasks
  /tasks/{id}/assignments/{user_id}:
    delete:
      description: Remove a user's assignment from a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove task assignment
      tags:
      - tasks
...
swagger: "2.0"